TRR flatfield source test (#279)
* Start integration of flat-field source characterization test into TRR sub-module.

* Continue integration of flat-field source characterization test into TRR sub-module.

* Small isort fix

* Fix path for matplotlib style file

* Account for flasher variance by default

* Streamline arguments

* Change default charge extraction method from GlobalPeakWindowSum to LocalPeakWindowSum

* Change required arguments to not required and use default values instead

* Fix run_path

* Let the matplotlib style control plot styles

* Try to fix doc CI by removing dependency to `tabulate` module.

* Small harmonization of camel cases for list of tests

* Change default output directory following the agreed convention "$NECTARCHAIN_FIGURES/trr_camera_X/script_name/figure.png"

* Fix handling of pickle files, to display plots in the TRR GUI at the end of a test

* All outputs are now written in `output_dir` following the agreed convention

* Pass charge extraction method and arguments to photostatistics tool

* Cast `charge_hg` and `charge_hg_std` as `np.float64`

* Remove user script for flat-field source characterization, which is now in TRR sub-module

* Less verbose output

* Fix typo

* All outputs are now written in `output_dir`

+++ src/nectarchain/makers/component/photostatistic_algorithm.py
@@ -144,8 +144,8 @@ class PhotoStatisticAlgorithm(Component):
         ).T
         self._results.is_valid = mask
         log.info("Trying to write charges")
-        self._results.charge_hg = self.meanChargeHG
-        self._results.charge_hg_std = self.sigmaChargeHG
+        self._results.charge_hg = np.float64(self.meanChargeHG)
+        self._results.charge_hg_std = np.float64(self.sigmaChargeHG)
 
         figpath = kwargs.get("figpath", False)
         if figpath:
